refactor: empty_activity to main_activity
