+++ presentation/src/main/kotlin/team/sungbinland/sseukssak/activity/main/MainActivity.kt
@@ -5,7 +5,7 @@
  * Please see: https://github.com/sungbinland/sseukssak/blob/main/LICENSE.
  */
 
-package team.sungbinland.sseukssak.activity.empty
+package team.sungbinland.sseukssak.activity.main
 
 import android.os.Bundle
 import android.view.View
@@ -16,34 +16,30 @@ import androidx.navigation.ui.AppBarConfiguration
 import dagger.hilt.android.AndroidEntryPoint
 import team.sungbinland.sseukssak.R
 import team.sungbinland.sseukssak.base.BaseActivity
-import team.sungbinland.sseukssak.databinding.ActivityEmptyBinding
+import team.sungbinland.sseukssak.databinding.ActivityMainBinding
+import team.sungbinland.sseukssak.util.extensions.hide
 import team.sungbinland.sseukssak.util.extensions.repeatOnStarted
 
 @AndroidEntryPoint
-class EmptyActivity : BaseActivity<ActivityEmptyBinding, EmptyViewModel>(R.layout.activity_empty) {
-    override val vm: EmptyViewModel by viewModels()
+class MainActivity : BaseActivity<ActivityMainBinding, MainViewModel>(R.layout.activity_main) {
+    override val vm: MainViewModel by viewModels()
     private lateinit var navController: NavController
     private lateinit var appBarConfiguration: AppBarConfiguration
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        binding.lifecycleOwner = this
         binding.vm = vm
 
         val navHostFragment =
             supportFragmentManager.findFragmentById(R.id.container_main) as NavHostFragment
         navController = navHostFragment.navController
         appBarConfiguration = AppBarConfiguration(navController.graph, binding.drawerLayout)
-
         setToolbar()
-
         setDrawer()
-
         repeatOnStarted {
-            vm.eventFlow.collect { event -> handleEvent(event) }
+            vm.eventFlow.collect(::handleEvent)
         }
-
     }
 
     private fun handleEvent(event: Event) = when (event) {
@@ -54,18 +50,18 @@ class EmptyActivity : BaseActivity<ActivityEmptyBinding, EmptyViewModel>(R.layou
 
     private fun setToolbar() {
         navController.addOnDestinationChangedListener { controller, destination, arguments ->
-            binding.toolbarMain.visibility = View.VISIBLE
+            binding.toolbarMain.hide(false)
 
 //            val customToolbarFragmentSet = setOf(
 //              R.id.~~
 //            )
 
             when (destination.id) {
-                R.id.nav_sseukssak -> binding.toolbarTitle.text =
+                R.id.nav_sseukssak -> binding.tvTitle.text =
                     getString(R.string.label_sseukssak_list)
             }
 
-//            if (customToolbarFragmentSet.contains(destination.id)) binding.toolbarMain.visibility = View.GONE
+//            if (customToolbarFragmentSet.contains(destination.id)) binding.toolbarMain.hide(true)
         }
     }
 
@@ -78,9 +74,7 @@ class EmptyActivity : BaseActivity<ActivityEmptyBinding, EmptyViewModel>(R.layou
             binding.mainNavi.constraintProfile,
             binding.mainNavi.btnService,
             binding.mainNavi.btnQuestion
-        ).forEach { viewId ->
-            viewId.setOnClickListener {
-
+        ).forEach { viewId -> viewId.setOnClickListener {
                 drawerLayout.closeDrawers()
             }
         }
