--- presentation/src/main/res/layout/activity_empty.xml
@@ -1,155 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ SseukSsak © 2022 Team Sungbinland. all rights reserved.
-  ~ SseukSsak license is under the MIT.
-  ~
-  ~ Please see: https://github.com/sungbinland/sseukssak/blob/main/LICENSE.
-  -->
-
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="vm"
-            type="team.sungbinland.sseukssak.activity.empty.EmptyViewModel" />
-
-    </data>
-
-    <androidx.drawerlayout.widget.DrawerLayout
-        android:id="@+id/drawer_layout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:fitsSystemWindows="true">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-            <androidx.appcompat.widget.Toolbar
-                android:id="@+id/toolbar_main"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent">
-
-                <androidx.constraintlayout.widget.ConstraintLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content">
-
-                    <ImageView
-                        android:id="@+id/toolbar_icon"
-                        android:layout_width="28dp"
-                        android:layout_height="28dp"
-                        android:layout_marginTop="20dp"
-                        android:contentDescription="@null"
-                        android:scaleType="centerCrop"
-                        android:src="@drawable/sungbinland"
-                        app:layout_constraintStart_toStartOf="parent"
-                        app:layout_constraintTop_toTopOf="parent" />
-
-                    <TextView
-                        android:id="@+id/toolbar_title"
-                        style="@style/TextMedium.Bold"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_marginStart="10dp"
-                        app:layout_constraintBottom_toBottomOf="@id/toolbar_icon"
-                        app:layout_constraintStart_toEndOf="@id/toolbar_icon"
-                        app:layout_constraintTop_toTopOf="@id/toolbar_icon"
-                        tools:text="쓱싹리스트" />
-
-                    <androidx.appcompat.widget.AppCompatButton
-                        android:id="@+id/toolbar_arrow"
-                        android:layout_width="24dp"
-                        android:layout_height="24dp"
-                        android:layout_marginStart="-7dp"
-                        android:layout_marginTop="-12dp"
-                        android:background="@drawable/ic_baseline_keyboard_arrow_down_24"
-                        android:onClick="@{() -> vm.openDrawer()}"
-                        app:layout_constraintStart_toEndOf="@id/toolbar_title"
-                        app:layout_constraintTop_toTopOf="@id/toolbar_title" />
-
-                    <TextView
-                        android:id="@+id/tv_total_prior"
-                        style="@style/TextSmall12"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:text="@string/label_total_prior"
-                        app:layout_constraintBottom_toBottomOf="@id/toolbar_title"
-                        app:layout_constraintEnd_toStartOf="@id/tv_total"
-                        app:layout_constraintTop_toTopOf="@id/toolbar_title" />
-
-                    <TextView
-                        android:id="@+id/tv_total"
-                        style="@style/TextSmall12.Bold"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        app:layout_constraintBottom_toBottomOf="@id/toolbar_title"
-                        app:layout_constraintEnd_toStartOf="@id/tv_total_rear"
-                        app:layout_constraintTop_toTopOf="@id/toolbar_title"
-                        tools:text=" 433,400개" />
-
-                    <TextView
-                        android:id="@+id/tv_total_rear"
-                        style="@style/TextSmall12"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:text="@string/label_total_rear"
-                        app:layout_constraintBottom_toBottomOf="@id/toolbar_title"
-                        app:layout_constraintEnd_toEndOf="@id/btn_search"
-                        app:layout_constraintTop_toTopOf="@id/toolbar_title" />
-
-                    <EditText
-                        android:id="@+id/btn_search"
-                        style="@style/SearchEditText"
-                        android:layout_width="0dp"
-                        android:layout_height="37dp"
-                        android:layout_marginTop="20dp"
-                        android:layout_marginEnd="20dp"
-                        android:background="@drawable/background_navy_grey_r24"
-                        android:ellipsize="end"
-                        android:gravity="start|center_vertical"
-                        android:hint="@string/label_search"
-                        android:paddingStart="20dp"
-                        android:singleLine="true"
-                        app:layout_constraintEnd_toEndOf="parent"
-                        app:layout_constraintStart_toStartOf="parent"
-                        app:layout_constraintTop_toBottomOf="@id/toolbar_icon" />
-
-                    <ImageView
-                        android:layout_width="20dp"
-                        android:layout_height="20dp"
-                        android:layout_marginEnd="10dp"
-                        android:elevation="2dp"
-                        android:src="@drawable/ic_baseline_search_navy_20"
-                        app:layout_constraintBottom_toBottomOf="@id/btn_search"
-                        app:layout_constraintEnd_toEndOf="@id/btn_search"
-                        app:layout_constraintTop_toTopOf="@id/btn_search" />
-                </androidx.constraintlayout.widget.ConstraintLayout>
-
-            </androidx.appcompat.widget.Toolbar>
-
-            <androidx.fragment.app.FragmentContainerView
-                android:id="@+id/container_main"
-                android:name="androidx.navigation.fragment.NavHostFragment"
-                android:layout_width="0dp"
-                android:layout_height="0dp"
-                app:defaultNavHost="true"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/toolbar_main"
-                app:navGraph="@navigation/navigation_graph" />
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-        <include
-            android:id="@+id/main_navi"
-            layout="@layout/activity_drawer" />
-
-    </androidx.drawerlayout.widget.DrawerLayout>
-
-</layout>
